% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignmentSeed.R
\name{assignmentSeed}
\alias{assignmentSeed}
\title{Use id, seed, and salt to create an integer seed}
\usage{
assignmentSeed(id, seed, salt)
}
\arguments{
\item{id}{id number of the student}

\item{seed}{assignment seed}

\item{salt}{assignment salt}
}
\value{

}
\description{
Use id, seed, and salt to create an integer seed
}
