% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_all.R
\name{compileAll}
\alias{compileAll}
\title{Compile a data set and assignment document}
\usage{
compileAll(
  assignment_data,
  id,
  seed,
  solutions,
  e = new.env(),
  type = "solve",
  pdf = FALSE
)
}
\arguments{
\item{assignment_data}{data from _assignment.yml}

\item{id}{student id}

\item{seed}{random seed}

\item{solutions}{show solutions?}

\item{e}{environment in which to compile the document}

\item{type}{source (for the file name)}

\item{pdf}{create PDF instead of HTML?}
}
\value{

}
\description{
Compile a data set and assignment document
}
