% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAssignmentHtmlDefault.R
\name{compileAssignmentHtmlDefault}
\alias{compileAssignmentHtmlDefault}
\title{A fallback function to compile an assignment, if none is given in the assignment YAML}
\usage{
compileAssignmentHtmlDefault(path, envir = new.env(), ...)
}
\arguments{
\item{path}{The path to the assignment folder}

\item{envir}{An environment in which to compile the assignment}

\item{...}{Other arguments (to be ignored)}
}
\value{
Path to the file containing the compiled HTML content (as a fragment, no header)
}
\description{
A fallback function to compile an assignment, if none is given in the assignment YAML
}
