% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilePage.R
\name{compilePage}
\alias{compilePage}
\title{Compile a data set and assignment document}
\usage{
compilePage(assignment_data, id, seed, solutions, e, init_list = list())
}
\arguments{
\item{assignment_data}{data from _assignment.yml}

\item{id}{student id}

\item{seed}{random seed}

\item{solutions}{show solutions?}

\item{e}{environment in which to compile the document}

\item{init_list}{assignment initialization information}
}
\value{

}
\description{
Compile a data set and assignment document
}
