% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileStatic.R
\name{compileStatic}
\alias{compileStatic}
\title{Generate a static HTML version of an assignment (without any shiny interaction).}
\usage{
compileStatic(
  id,
  assignment,
  masterseed,
  solutions = FALSE,
  entry = "download",
  format = "SPSS",
  plain = FALSE,
  ...
)
}
\arguments{
\item{id}{Student id (as would be entered in shiny app)}

\item{assignment}{assignment short code}

\item{masterseed}{master seed (as would be entered in shiny app)}

\item{solutions}{(boolean; default: `FALSE`) render the solutions?}

\item{entry}{mode to use (`solve` vs `download` currently)}

\item{format}{data format (for the buttons)}

\item{plain}{if `TRUE`, do not use flexDashboard}

\item{...}{Additional arguments to `rmarkdown::render`}
}
\value{
The return value of `rmarkdown::render`
}
\description{
Intended for use with, e.g., exams, where no solutions will be offered. 
Using a static page dramatically reduces resources needed for deploying
an assignment that does not requre interactivity.
}
\details{
The static version uses the same seed code as the interactive shiny app,
so the compiled document will be consistent (as long as the ID and master seed are). 

Download buttons have embedded files saved as data URIs.
}
