% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignments.R
\name{getAssignments}
\alias{getAssignments}
\title{Get a list of assignments in a directory}
\usage{
getAssignments(pkg = pkg_options()$assignment_pkg, simple = TRUE)
}
\arguments{
\item{pkg}{Package from which to pull the assignments (character string)}

\item{simple}{return only a named vector (for selectInput)}
}
\value{

}
\description{
Get a list of assignments in a directory
}
