% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Run one of the shiny interfaces}
\usage{
launchApp(which = "solve", query = "")
}
\arguments{
\item{which}{Which app to open ('solve' or 'download')}

\item{query}{Query string to add to the URL}
}
\value{

}
\description{
Run one of the shiny interfaces
}
