% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_options.R
\name{pkg_options}
\alias{pkg_options}
\title{Set or get options for my package}
\usage{
pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{date_format_yaml}}{(\code{character}) Format from which to parse dates in YAML for assignments }
 \item{\code{date_format_prnt}}{(\code{character}) Format in which to print dates (e.g., in error messages) }
 \item{\code{initial_seed}}{(\code{character}) Default seed to put in the seed box, if no query string is used }
 \item{\code{start_practice_message}}{(\code{character}) Message to give when switching to practice mode }
 \item{\code{start_assignment_message}}{(\code{character}) Message to give when switching to assignment mode }
 \item{\code{secret_filename}}{(\code{character}) Filename that contains the secret (in the last line of the file) }
 \item{\code{assignment_pkg}}{(\code{character}) Installed package to use for the assignments }
}
}

