% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ran_alphanum.R
\name{ran_alphanum}
\alias{ran_alphanum}
\title{Create random alphanumeric strings}
\usage{
ran_alphanum(n = 1, n_char = 10)
}
\arguments{
\item{n}{number of strings}

\item{n_char}{length of strings}
}
\value{

}
\description{
Create random alphanumeric strings
}
