% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomAlphaNum.R
\name{randomAlphaNum}
\alias{randomAlphaNum}
\title{Create random alphanumeric strings}
\usage{
randomAlphaNum(n = 1, n_char = 10)
}
\arguments{
\item{n}{number of strings}

\item{n_char}{length of strings}
}
\value{

}
\description{
Create random alphanumeric strings
}
