% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceAll.R
\name{sourceAll}
\alias{sourceAll}
\title{Source all the files indicated in the assignment data}
\usage{
sourceAll(assignment_data, e)
}
\arguments{
\item{assignment_data}{data for the particular assignment}

\item{e}{environment in which to source}
}
\value{

}
\description{
Source all the files indicated in the assignment data
}
