% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssignments.R
\name{getAssignments}
\alias{getAssignments}
\title{Get a list of assignments in a directory}
\usage{
getAssignments(path = system.file("assignments", package = "flexTeaching"))
}
\arguments{
\item{path}{Directory path}
}
\value{

}
\description{
Get a list of assignments in a directory
}
