% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_options.R
\name{pkg_options}
\alias{pkg_options}
\title{Set or get options for my package}
\usage{
pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{date_format_yaml}}{(\code{character}) Format from which to parse dates in YAML for assignments }
 \item{\code{date_format_prnt}}{(\code{character}) Format in which to print dates (e.g., in error messages) }
}
}

